+++ ESFramework/Core/Additions/NSDate+ESAdditions.h
@@ -14,11 +14,6 @@
 
 @interface NSDate (ESAdditions)
 
-/**
- * Timestamp
- */
-+ (NSTimeInterval)timeIntervalSince1970;
-
 - (BOOL)isBefore:(NSDate *)aDate;
 - (BOOL)isAfter:(NSDate *)aDate;